Isolate delete and change actions
Isolate interaction with the html elements and editing the
Library #books array into separate methods

Leverage event delegation to capture all click event handlers on the
table body instead of duplicating event handlers on each btn

Remove 'delete' and 'change' html classes on the html btns. They are
no longer needed as the event handler is now referencing the
.textContent prop.

Note for future change:
Had to include a temp workaround to avoid duplicating event listeners
by using a two-step init approach. The issue is that the click
listeners are on the table body which is not removed by the
updateView call.