Refactor modal event listeners and functions
Separated the extraction of modal input element names and values
into it's own function. This function now returns the an object
of name: value pairs. Therefore, the new Book call was changed to
properly reference the object instead of an array index.

Also created a simple validation of html input elements to catch any
blank input values and fire an alert to the user. Will continue to
develop more validations in future commits.

Minor updates to the html element names to match Book props.